+++ doc/changelog.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 * :bug:`- major` Correctly call fixture.test_end on test errors (Thanks @pierreluctg!)
+* :feature:`-` Resuming sessions now supports ``--failed-only`` and ``--unstarted-only`` to control which tests to resume
 * :bug:`879 major` Fix handling exceptions during ``entering_debugger`` hook triggered by ``Application.__exit__()``
 * :bug:`878 major` Fix notification plugin's handling of backslash plugin with no active session
 * :release:`1.7.2 <11-10-2018>`